requirements.txt
streamlit
langchain
requests
python-dotenv
pandas
transformers
langchain_community
langchain-groq